CDAP-693 Adding the ability to run hive queries on streams. This includes various changes to make this possible.
Streams are now created when an application is deployed
instead of lazily created when the first event is written to it.
If explore is enabled when a stream is created, an external
hive table is created using a custom HiveStorageHandler for
streams. This required some refactoring around the stream input
format, as our mapreduce code uses the newer mapreduce apis, but
the hive code uses the old mapred apis.