Better stream events fetch logic.
The improvement is mainly for the case when a fetch request comes with
a startTime, and the first event with the given startTime is in higher
stream file sequence. The old logic will return no events, as the
live stream reader expects one to keep polling for events (when
switching to new file, it will return 0 events and expect caller to call
read again. This works well for Flow).

The new logic is simply to honor that behavior by keep
polling as long as there is events being read/skipped.