Fix JMS realtime source to pass the right parameters via the PluginConfig.
Add more required parameters for JMS JNDI naming via the JmsPluginConfig rather than
pass through from properties for easy testing.

It was to test before because the parent PluginConfig's properties is private
and the method to get the properties is final so need to push some of the java naming
properies as top leve in JmsPluginConfig class.