CDAP-2202 Validate RunRecord for running program agains ProgramRuntimeService as source of truth
Currently the call to check run status of a program is run against run records' store.
In case of abrupt shutdown, there is no chance to update the state of the run record of the programs.
For example if a Flow program is running in standalone then the programs dies, then the run record
for that program will indicate still running when the CDAP is restarted.

So, the call for
  http://localhost:10000/v3/namespaces/default/apps/CountRandom/flows/CountRandom/runs?status=running
and
  http://localhost:10000/v3/namespaces/default/apps/CountRandom/flows/CountRandom/status
could return different results.

The proposed fix to change the Adapter and Program lifecycle to check against the runtime service
when querying for run records of a program.

For Adapter we just need to check for worker one because the workflow is done via scheduler so
when cdap start the workflow adapter should start the scheduled programs accordingly.