CDAP-2490 Check if invalid Run Records for Spark and MapReduce part of run from Workflow child programs
Since 3.0 we have scheduled thread to update Run Record store for thos with RUNNING state but does not
have real program running via check to RuntimeService.

Howerver, for Run Records for Spark and MapReduce that are part of run from Workflow child programs,
the runtime service never return runtime info since those runs are run from Workflow driver.

Which means that there is no Twill application for the internal  MapReduce and Spark child programs run records.

Need to update the RunRecord checker to account for this situation.

This commit add check for run records that has workflowrunid properties which are children of Workflow.