CDAP-2208 fixing dataset and explore classloading issues
The first issue is that the FilterClassLoader does not filter
when getResource is called. When a dataset jar is built, this
can result in system jars getting leaked into the dataset jar.

The second issue is that in explore, the classloader used to
instantiate a dataset a DirectoryClassLoader using the dataset
jar and with the system classloader as a parent. This means
that system classes are leaked. Instead, the parent classloader
must be a FilterClassLoader. To support this, adding a couple
new ways to create a FilterClassLoader.