CDAP-2564 refactoring RemoteDatasetFramework so that it no longer unpacks dataset jars by default each time it instantiates a dataset or dataset type. This is because in most contexts, the classloader passed in will be a ProgramClassLoader that already has access to the dataset.
For calls that are made in the CDAP system context, added a
separate method to DatasetFramework that takes a
DatasetClassLoaderProvider that can unpack
dataset jars if it needs to, but which is closeable to allow
cleanup of those unpacked jars. This method is used by the dataset
admin op handler and by explore to instantiate datasets.

Also refactoring the way datasets are instantiated in explore,
since there is no longer any need to cache dataset classloaders,
but need to make sure to clean up unpacked directories. To do this,
adding a SystemDatasetInstantiator class that is intended to be
used in CDAP system contexts, which will delete relevant
local directories when it is closed. This replaces most direct
interactions with DatasetFramework.