CDAP-2564 make directory classloader closes its classloader too
The DirectoryClassLoaderProvider will now close its cached
classloaders. This is needed since URLClassLoader needs to be
closed so that it doesn't hold on to file descriptors. Also
fixing a bug where dataset module classloaders were not being
chained correctly. For datasets that use other custom datasets,
this would have been a problem.