CDAP-2665 add artifact repository
The ArtifactRepository is a layer on top of the ArtifactStore
and is responsible for inspecting artifacts to determine the
metadata for the artifact before writing it to the store.
This will eventually replace the PluginRepository, though there
is some more refactoring required before that is possible.
For now, extracting jar inspection logic to a common
ArtifactInspector class that both repositories use, and refactoring
plugin instantiation and classloading so that it can be used by
both repositories.

Changing the API of the ArtifactStore a little so that it
returns sorted maps so that it is easier to select which
artifact to use for a specific plugin.