CDAP-2530 making RecordScannable<StructuredRecord> explorable
There was a bug where Table and ObjectMappedTable are
explorable because they are RecordScannable<StructureRecord>,
but custom datasets were not explorable in the same way. So
somebody could create a custom dataset that uses an explorable
table, but would not be able to use that to make their custom
dataset explorable.

This fix will create a Hive table for custom datasets that are
RecordScannable<StructuredRecord>, assuming the schema dataset
property is set. If the schema property is not set, enabling
exploration will fail with a message that the schema property
is required.