CDAP-3042 fix bug where buffered operations weren't seen in multiget
Fixes a bug in BufferingTable where multiget calls for HBase were
only returning persisted data and could not see changes made in the
same transaction. In order to fix this, BufferingTable now has a
getPersisted(List<Get> gets) method for subclasses to override.
BufferingTable will then apply buffered changes on top of the
persisted changes.

Also adding a service endpoint to the UserProfiles example app to
demonstrate usage of multi-get on a Table.