CDAP-3042 remove always false condition in BufferingTable
In the get code paths, there was a check for deleted rows by
checking if the buffer contains a key but has a null value for
the key. But the way deletes are implemented, they fetch all
columns for the key and set the column values to empty bytes,
so the condition can never be satisfied. Removing the check.