[CDAP-1983] Added an application for processing and analyzing Wikipedia using a CDAP Workflow.
The app contains a CDAP Workflow that runs in either online or offline mode. In the offline mode, it expects Wikipedia data to be
available in a Stream. In the online mode, it attempts to download wikipedia data for a provided set of page titles (the output of
the Facebook Likes API). Once wikipedia data is available it runs a map-only job to filter/normalize data formatted as text/wiki-text
into text/plain. It then runs two analyses on it in a fork -
1. A Spark program that does topic modeling using Latent Dirichlet Allocation (LDA).
2. A MapReduce program that does a simple 'Top N' word generation.

This app also demonstrates Workflow Token usage in:
1. Condition Predicates
2. Determining Job configuration (setting it based on values in the token)
3. map/reduce functions
4. Spark Programs

It also shows usage of getDataset + put() in Spark programs.