CDAP-2768 filter out system.queue from list of datasets
system.queue is a system dataset, but it exists in each namespace
once a flow has been started in the namespace. It should not be
visible via the get datasets rest call. This is a short term
fix, with the longer term fix to have support for system datasets
in namespaces other than the system namespace.