CDAP-3104 enabling explore on a Table without schema should no-op
Fixes a bug where a request to enable exploration on a Table
without a schema would error. This is required because a call to
enable exploration is made on every single dataset when it is
created or when its properties change. All this dataset specific
logic in explore should be moved out, but that is outside the
scope of this fix and has its own open issue.