CDAP-2764 use artifact repo in app deployment pipeline
Some refactoring to make the application deployment pipeline add
the jar to the artifact repo, then call the deployment pipeline
with the id for the newly added artifact. Unfortunately, the
old code path still exists because app templates rely on it, but
should be removed soon. ArtifactID is now also part of the app
spec, which is a requirement for plugin integration with apps.