CDAP-3261 fixes for adding artifacts in distributed mode
Removing location serialization and deserialization through a gson
type adapter since it triggers an infinite loop in gson for some
reason. Also changing add artifact api a little bit to take an
input supplier for an input stream instead of just the input
stream to ensure that the stream is closed.