CDAP-3319 make sure artifacts cant have grandparents
This is already a limitation of the plugin system, but it was not
being enforced by the artifact framework. We want to keep things
simple so that an artifact can extend another artifact, but only
to one level. This makes it simpler for our framework, but also
prevents users from creating complicated dependency trees that
are hard for them to manage as well.