CDAP-3283 index application class information and expose it
adding rows to the ArtifactStore metatable for application classes
found in artifacts to allow efficient lookup of app classes in
a namespace and for a given class name. This functionality is
exposed through the ArtifactRepository and through the http
handler.

Filling in the REST endpoint for getting application classes in
a namespace and by classname. Also adding artifact name and version
filters to the get apps REST endpoint.

Finally, also added methods to ArtifactClient corresponding to the
new REST APIs, and a new method to ApplicationClient for filtering
applications by artifact.