CDAP-3450 move app deploy logic from http handler into service
Some refactoring to move logic out of the AppLifecycleHttpHandler
into AppLifecycleService. The handler no longer uses Manager for
app deployment or ArtifactRepository for managing artifacts.
Instead, the logic was moved into the service where it can be
re-used. For example, CDAP upgrade may need to make use of this
logic when upgrading existing apps to the new artifact framework.