CDAP-3539 fixing upgrade to handle bad app metadata
Fixing an issue where the upgrade tool would choke on apps that
have the wrong location stored for the application archive. Not
sure how that can happen, it is probably a sign on multiple
cdap upgrades and one of those upgrades going wrong.

In any case, check the location exists. If not, try to find it
at the expected location. If that doesn't exist, just continue
on and log a message.