CDAP-3560 change isSystem to a scope enum
Changing the isSystem boolean to an enum for artifact scope.
We will start out with system and user scopes, but this allows us
to add new scopes in the future and is consistent with other
concepts like metrics.

Also doing some cleanup so that ArtifactId is used in a few
classes instead of storing the individual id fields.

Also adding missing package-info file for artifacts.