CDAP-3490 move deletion of templates to AdapterService
Instead of checking the application name to see if an app is an
application template, moving template deletion logic into the
adapter service. Adapter service can tell what is a template
because the presence of an adapter means there is a deployed
application template. And if there are no adapters, we know the
corresponding template was already deleted.