CDAP-3582 allow filtering apps by multiple artifact names
Extending the get apps rest api to allow specifying a comma
separated list of artifact names. The endpoint will return all
apps that use an artifact whose name is in that set of names.

Also performing some refactoring along the way, moving logic from
the http handler into the service. This allows us to remove the
usage of Store in the http handler.