CDAP-3591 ability to write to multiple sinks in etl realtime
Adding the ability to write to multiple sinks in the etl realtime
application. The exact same data will be written to all sinks.
If transactional sinks are used, a failure to write to one sink
will cause that write to other sinks to fail. If non-transactional
sinks are used, a write to one sink can succeed while a write to
another sink can fail.