CDAP-3592 refactoring to prepare for multiple sinks in etl batch.
Switching ETLMapper to implement ProgramLifecycle<MapReduceTaskContext>
instead of ProgramLifecycle<MapReduceContext>. This will give access
to the new context.write(name, key, value) methods that will be
required to support multiple sinks.

In order to do this, introduced a BatchRuntimeContext class that is
passed to sources, transforms, and sinks at initialize time instead of
passing BatchContext. This is because BatchContext allows you to get
the Hadoop Job, add outputs, and set input - all of which really only
makes sense in the prepareRun (beforeSubmit) method. In addition,
MapReduceTaskContext has no way to get the Hadoop Job, so the change
is required. In the process also cleaning up the context classes to
remove duplicate code and remove usage of AdapterContext.