CDAP-3727 fix bug with order of system artifact loading
The previous implementation tried to enforce ordering by sorting
on the artifacts, with parents ordered larger than children. The
problem was the comparison make sense when 2 artifacts were being
compared, but it wasn't a natural ordering. That is, there could
be a scenario where A < B, B < C, but A > C.

Instead, since we enforce only 1 level of dependencies, just
adding all parents first, then adding the children.