CDAP-3860 fixed bugs with available plugins for an artifact
Fixed a bug where plugins in an artifact were not getting included
in the list of plugins available to that artifact. In other words,
only plugins from child artifacts were being included.

Also fixed a bug where listing the plugins for a non-existent
artifact would return an empty list instead of a 404.