CDAP-3939 improve error handling with artifact and app deployment
Improving the failure handling for the API call that creates an
artifact and application in the same call. If the artifact was
added successfully, but the application failed to deploy,
remove the newly added artifact so that deployment can be
re-tried. If the artifact already exists, give a more detailed
error message explaining how to deploy the app.