CDAP-4147 make app update test more robust.
The test fails on slow machines because artifact deployment takes
longer than the client read timeout. Fixing the test so that the
deployment exception shows up in the logs instead of the cleanup
exception. Also increasing the read timeout and fixing a bug in
the ClientConfig builder where read and connect timeouts were
getting swapped.