CDAP-3694 fix examples poms
Exclude asm from the common-http dependency. Without this,
an old version of asm (4.0) gets pulled in instead of the version
the unit-test module requires, causing unit tests to break.
This was causing PurchaseHistory and Wikipedia example tests
to fail. The test failures are only seen when using the sdk.
In the cdap project, the correct version of asm is used.

Also excluding netty and jetty dependencies from StreamConversion
example to fix test failures.