CDAP-4400 Adding experimental support for ETLSpark.
Adds an engine config setting to batch pipelines that lets users
configure whether to use mapreduce or spark. Spark support is
experimental, as there are several issues with it in its current
form.

Lookup and error datasets do not work yet, and the destroy
method of plugins is never called, which means any plugins that
rely on destroy will not release their resources.
Data will also be reprocessed when writing to multiple sinks.