CDAP-3945 change to limit schedule runs instead of program runs
Changing the constraint to be based off schedule runs instead of
program runs. As part of this, added system arguments to the
metadata of a run record in order to be able to tell which run
records were started by a schedule versus started manually.

Also changing exceeds to setting a max since that seems clearer.