CDAP-4226 enhance build to optionally copy additional artifacts
Adding a property that will allow the build to copy additional
artifacts into the artifacts directory. This can be used to
include artifacts from other projects that we want to be included
as system artifacts. For example, we plan on including various
hydrator plugins that may come from different repositories.

Also simplifying etl lib build to only create a single jar, since
the three it was creating before were exactly the same except for
their manifests.