CDAP-4228 add ability to select plugin artifact in etl apps
Adding configuration settings to the etl apps that will select
a plugin's artifact using scope, name, and/or version settings
provided by the user. The app will choose the first artifact
that matches all of the given artifact settings. If scope,
name, or version is not given, the app will not try to match
that field. If no match is found, an exception is thrown and the
app is not created.