CDAP-4225 added an upgrade tool for Hydrator
The 3.3.x versions of the etl applications are backwards
incompatible with the 3.2.x versions. The upgrade tool offers
a way to upgrade all pipelines, all pipelines in a namespace,
or a specific pipeline. Also offers a way to take a request that
could have been used to create a 3.2.x pipeline and convert it
into a request to create a 3.3.x pipeline.