Switch  the key/value pairs stored for schema to make searching for a field (with no type specified easier).
This also adds URL encoding/decoding for the 'query' query param. However, this should not be done, because this way, users are expected to encode the param while making requests. We should instead accept a separate 'context' or 'type' parameter for the search API that specifies 'field' to search for fields, and translate it into the right query internally, instead of exposing the internal details to users.
In future, we should also add an API that retrieves all available search contexts for a given entity.