CDAP-4233 inserting stage name in log messages from etl stages
The ETL applications will now modify the appenders for the root
logger so that log event messages will be prefixed by the name
of the stage they were logged from.  Uses static MDC methods
to communicate the stage at the time of logging. Most of the code
revolves around setting the stage in MDC when plugin code is
running, and removing the stage when application or cdap platform
code is running.