CDAP-4539 only set hive classpath for master service
Fixing init scripts so that only the master service tries to set
the explore classpath, instead of all cdap services.
Also tweaking the set hive classpath function to print the
contents of stderr for the hive command when it fails so that the
user has some idea of what went wrong.