CDAP-4536 check infrastructure on master startup
adding some sanity checks on master startup to make sure hdfs,
yarn, and hbase are available before starting up the master.
If any are not, master will shut down.

which checks are run can be configured in the cdap config
by either specifying the packages that contain the checks,
or by specifying exact classnames. There is also a setting
to turn off all checks.