CDAP-4566 separate startup checks and run by init scripts
Moving startup checks from MasterServiceMain to its own class so
that they are run in a separate command. This is so we can call
it separately in the init script. This will make it so that a
startup failure will cause the init scripts to exit with a failure,
rather than having the scripts exist with a success, then have the
master fail soon after.