CDAP-4581 protect against root directory in url classloader
fixing a problem with the startup check runner when the classpath
contains empty entries and the tool is run from the root directory.
In those cases, the working directory, which is the root directory,
gets added as a uri for the url classloader, which will cause
ClassPath to scan the entire filesystem.