CDAP-4294 namespace isolation for system artifact plugins
Fix a bug where users in one namespace could see plugins from
another namespace when using a system application artifact.
To fix this, when deploying an app, we need to pass down the
namespace the app will be in to filter out plugin artifacts from
other namespaces. Most of the changes revolve around passing
namespace to the configurers that require it.