CDAP-4723 add build improvements to plugin archetypes
add the build script used to create the deployment json for etl
plugins. This change will create the json file in the target
directory instead of forcing users to create it themselves. It
also introduces the docs and widgets directories that integrate
with the Hydrator UI to format plugin fields and provide reference
documentation.

Also add a readme that describes how to build the plugin,
how to integrate with the UI, and how to deploy the plugin.