CDAP-4800 handle edge case for version range
Error when a version range has a lower version that is equal to
the upper version, but the lower version is inclusive and the
upper is exclusive. Also fixing a unit test that was trying to
test for this scenario, but calling the wrong method.

Also a small change to the artifact config parser so that the
right error message is used in the above scenario.