Fixes saving default version of a plugin based on user choice
Until today we have been storing just the version of a plugin keyed by 'pluginname-plugintype-artifactname'.
This will be a problem when we want USER scoped artifact and SYSTEM scoped artifact plugins to be surfaced in UI.
When the user chooses a USER scoped plugin we should make sure we request for USER scoped plugin's widgetJSON and not SYSTEM
scoped plugin's widgetJSON. To avoid this we record the entire artifact object instead of just the version when the user
chooses a plugin in UI in the user store.