CDAP-4944 make hydrator upgrade more robust
Fix a bug when the upgrade tool is told to only upgrade pipelines
but still tries to upgrade ui data.

The tool assumes that the ui data is in the 3.2.x format. This may
not be true, as the console store is not cleaned up or upgraded
in between cdap versions. This change moves ui data upgrade to the
end so that it does not interfere with upgrading pipelines, which
is more important. It also does not assume application names are
cdap-etl-batch or cdap-etl-realtime, since that is not the case
with really old ui data.

In general... that console store needs to be removed.