CDAP-4967 fix a schema parsing bug due to repeat inner records
The bug occurs when a record has a field of type record X,
and another field that uses record X as an inner field.
In such cases, the inner X is just a reference to the record name
instead of a full schema, which causes problems when trying to
create the schema for that field. The fix is to keep a global map
of all known record types instead of a map per record.