CDAP-5094 always use explore schema if it exists
Fixed a bug where the explore schema was not being used unless
the explore format was also being set. The explore schema should
always be used. It just happened to be the case that avro filesets
work with Hive without an explore schema because the schema can
be set through the avro.schema.literal table property.