CDAP-4772 remove deprecated app create restful api
When we first introduced artifacts, we overloaded the
PUT /apps/<app-id> api to have different behavior based on the
Content-Type header. This was to keep the endpoint backwards
compatible.

The old behavior, which expected the request body to be the jar
contents, has been deprecated for several releases so we are
removing it.