CDAP-5180 add conversions between config versions
Moves the configs from 3.2.x from the upgrade tool into the proto
module. Also moves the upgrade logic from them and formalizes the
upgrade process so that each config knows how to upgrade to the
next version, and upgrades can be chained together to reach the
final version.

Adding upgrade logic from v1 of the configs to v2, which will
allow us to internally use v2 everywhere when we implement the
upcoming workflow planner.