CDAP-5186 add dag class to prepare for hydrator planner
Extracts some of the logic in the
PipelineRegisterer into a Dag class which can be used to check
for cycles, linearize, etc. This class will be enhanced next to
include methods to help with pipeline planning.

Also uses the Dag class to validate the user provided ETLConfig and
generate a PipelineSpec.  The spec contains all information needed
to plan the pipeline execution.