CDAP-5186 add planner dags
Also adds a ConnectorDag class that helps with splitting a dag into
multiple dags based on special knowledge about certain types of
nodes. The ConnectorDag takes a normal dag as input as well as a
set of reduce nodes. Reduce nodes basically map to plugins in the
pipeline that would require a reduce step in the mapreduce program.
The ConnectorDag uses this information to mark certain nodes in the
dag as connector nodes, which is a node that the dag can be split
on to create multiple sub-dags. Each connector node will translate
into a local dataset source and sink that will be used as temporary
storage linking mapreduce programs. This is used by the PipelinePlanner
to create one or more phases to be executed in the workflow.

Also adds a ControlDag class that is used to transform an arbitrary
data flow type of dag into a fork-join dag usable by workflow. It
moves phases around to accomplish this, while maintaining all
happens-before relationships. It can then be used by a workflow at
configure time.