CDAP-5186 add connector implementations
Adds a source and sink for connector nodes. Also introduces a
PipelinePluginInstantiator that will instantiate these sources
and sinks directly instead of the through the plugin context, and
will instantiate all other plugins through the plugin context.
This is because we dont want to expose these classes as actual
sources and sinks that people can use.

As part of this work, refactoring PipelinePhase so that it tracks
plugin types instead of having specific knowledge about what a
source or transform or aggregator is.