CDAP-5146 refactoring and cleanup to prepare for new plugin types
ETLMapReduce is in dire need of cleanup so that we can more easily
modify it in the future. This takes a pass
at removing unneccessary logic, consolidating common logic,
removing warnings, and other general cleanup.

Consolidated onRunFinish logic that was happening in both spark
and mapreduce, which now both use the CompositeFinisher to avoid
special logic for sources and special logic for sinks and for
maintaining state about source, sinks, and their contexts.

Consolidated logic for creating a TransformExecutor into the
TransformExecutorFactory class, which should also help a good
deal when we need to create an executor in both the mapper and the
reducer.