CDAP-5186 implement aggregator plugin type
Implementing the aggregator plugin type for mapreduce engine.
Each aggregator translates to a reduce step in a mapreduce job.
The group by method is called in the mapper in order to decide
what to reduce on. The aggregate method is called first thing in
the reducer.

Moved most of the logic formerly in ETLMapper into TransformRunner
so that it can be re-used in the new added ETLReducer. Also moved
OutputWriter classes that were internal to ETLMapReduce out of the
class so that they can be re-used in the TransformRunner.