CDAP-5186 make new etl config backwards compatible
Adds the old v1 fields (source, transforms, sinks) to the v2
configs, and add a method to convert those fields into the
stages field.

With that, we can switch the apps to the v2 configs and remove
the old PipelineRegisterer logic that has been replaced with the
PipelineSpecGenerator. This also makes it so that the apps can
now understand the new style configs, which sets up implementation
of the smart workflow and aggregator plugin types.