fix connector dag to allow connectors in front of a sink
Fixed a bug in the connector dag that prevented sinks from having
connectors in front of them. We need a connector in front of a
sink if the sink is being written to by multiple sources. Without
this, multiple mapreduce jobs can try to write to the same sink
at the same time, which will cause issues for certain sinks.