CDAP-5146 implement aggregators in spark program
Added an implementation for aggregators in spark. Implementation
is not ideal for a few reasons. The first is that we can't run
multiple spark programs at the same time in a workflow, so we
are forced to squash a parallel jobs into sequential jobs.
The second is that we should be able to compute the entire pipeline
in a single spark job rather than splitting it across multiple.

As part of this change, had to make StructuredRecord an Schema
implement the Serializable interface in order for them to be
usable in Spark.

Also made a small fix so that enums in application configs dont
have to be given as all caps.