add planner logic for isolation nodes
An isolation node is one that should be bracketed on both sides
by connectors. This is to support the spark transform plugin type,
which will read from a source, perform some rdd operation, then
write to a sink.  If inserted into the middle of a pipeline, we
will need to write to a local dataset, then use that local dataset
as input into the spark transform.