CDAP-5512 plugin support in workflows
Add plugin support in workflows. This amounts to extending the
WorkflowContext interface so that it also implements PluginContext,
and extending WorkflowConfigurer so that it also implements
PluginConfigurer, and extending AbstractWorkflow so that it extends
AbstractPluginConfigurable.

The rest is just re-using the plugin work already done for the
other program types to pass a PluginInstantiator into
BasicWorkflowContext.