CDAP-5605 fix incorrect aggregator metrics
Fixed a bug where the metrics for records in and out of an
aggregator were incorrect because they were counting intermediate
group metrics.

Includes some refactoring so that metrics are all handled in
TrackedTransform instead of partly in TrackedTransform and partly
in DefaultEmitter. Also making the tracking classes a little more
generic so that the metric emitted can be set. This allows us to
introduce a new 'aggregator.groups' metric to count the number of
groups in an aggregator.