CDAP-5786 fix hydrator etl-realtime upgrade
Fixed a bug in the upgrade logic that would fail if certain optional
settings -- like instances -- were not set.

Also making general improvements to the upgrade tool.
Logging was not set up correctly, so configuring error and warn
messages to write to stderr and everything else to stdout. Also
adding an option to write configs that failed to upgrade to an
error directory for manual investigation. Pipelines can fail to
upgrade if they use plugins that are not backwards compatible.