Adds Redux, Redux-Thunk & used in rewriting left panel store in Hydrator++.
The left panel store is re-written to use Redux to be implemented in a more functional way.
This way we have a more functional approach with unit tests.

This commit rewrites LeftPanelStore as a map of plugin types and list of extensions. This is exactly how we show
things in left panel. Extensions include "batchsource", "batchsink" etc., and plugin types map
maintains a map of each extension to plugins (mapping each plugin with a list of all artifact versions supported by backend).