CDAP-5676 bump up hive version to pick up bug fixes
In particular, to ensure that partquet-1.6 is being used instead
of a parquet release candidate. This includes the fix for
PARQUET-161, which causes errors in explore for parquet schemas
with lots of null values.

This change only affects the hive version used by the sdk. In
distributed mode, the hive version and dependencies on the cluster
are still used.